+++ vite-project/src/config/api.jsx
@@ -1,5 +1,6 @@
 import axios from "axios";
-import {API_URL} from "./constants.js";
+import {API_URL} from "./constants.jsx";
+import {useNavigate} from "react-router-dom";
 
 const api = axios.create({
     baseURL: API_URL,
@@ -8,9 +9,9 @@ const api = axios.create({
 // Add a request interceptor
 api.interceptors.request.use(
     (config) => {
-        const access_token = localStorage.getItem('access_token');
-        if (access_token) {
-            config.headers.Authorization = `Bearer ${access_token}`;
+        const token = localStorage.getItem('token');
+        if (token) {
+            config.headers.Authorization = `Bearer ${token}`;
         }
         return config;
     },
@@ -22,29 +23,29 @@ api.interceptors.response.use(
     (response) => response,
     async (error) => {
         const originalRequest = error.config;
+        const navigate = useNavigate();
 
         // If the error status is 401 and there is no originalRequest._retry flag,
-        // it means the access_token has expired and we need to refresh it
+        // it means the token has expired and we need to refresh it
         if (error.response.status === 401 && !originalRequest._retry) {
             originalRequest._retry = true;
 
             try {
-                const old_refresh_token = localStorage.getItem('refresh_token');
                 const response = await axios.post(`${API_URL}/users/tokens/refresh`, {}, {
                     headers: {
-                        Authorization: `Bearer ${old_refresh_token}`
+                        Authorization: `Bearer ${localStorage.getItem('refresh_token')}`
                     }
                 });
-                const { token, refresh_token } = response.data;
+                const {token, refresh_token} = response.data;
 
-                localStorage.setItem('access_token', token)
+                localStorage.setItem('token', token)
                 localStorage.setItem('refresh_token', refresh_token)
 
-                // Retry the original request with the new access_token
+                // Retry the original request with the new token
                 originalRequest.headers.Authorization = `Bearer ${token}`;
                 return axios(originalRequest);
             } catch (error) {
-                // Handle refresh access_token error or redirect to login
+                return navigate('/login')
             }
         }
