+++ vite-project/src/config/constants.jsx
