+++ vite-project/src/pages/posts/SinglePostPage.jsx
@@ -1,5 +1,5 @@
 import axios from "axios";
-import {API_URL, TEST_API_URL} from "../constants.js";
+import {API_URL, TEST_API_URL} from "../../config/constants.jsx";
 import {useLoaderData} from "react-router-dom";
 import { CloudArrowUpIcon, LockClosedIcon, ServerIcon } from '@heroicons/react/20/solid'
